added following fields to create child
race
date_of_birth
registration_latitude
registration_longitude
citizenship
gender
