+++ src/constants/index.js
@@ -9,6 +9,7 @@ export * from './icons'
 export * from './colors'
 export * from './layout'
 export * from './requests'
+export * from './meta'
 
 export const APP_VERSION = '2.0.0-alpha' // Native App Version
 export const BUNDLE_VERSION = 1 // JS Bundle Version - internal usage only (code-push)
