+++ src/libs/Utils.js
@@ -16,240 +16,5 @@ export default class Utils {
     return toRet
   }
 
-  static getCitizenships = () => {
-    let tmp = {
-      'ZA': 'South African',
-      'AF': 'Afghanistan',
-      'AX': 'Aland Islands',
-      'AL': 'Albania',
-      'DZ': 'Algeria',
-      'AS': 'American Samoa',
-      'AD': 'Andorra',
-      'AO': 'Angola',
-      'AI': 'Anguilla',
-      'AG': 'Antigua and Barbuda',
-      'AR': 'Argentina',
-      'AM': 'Armenia',
-      'AW': 'Aruba',
-      'AU': 'Australia',
-      'AT': 'Austria',
-      'AZ': 'Azerbaijan',
-      'SH': 'Ascension',
-      'BS': 'Bahamas',
-      'BH': 'Bahrain',
-      'BD': 'Bangladesh',
-      'BB': 'Barbados',
-      'BY': 'Belarus',
-      'BE': 'Belgium',
-      'BZ': 'Belize',
-      'BJ': 'Benin',
-      'BM': 'Bermuda',
-      'BT': 'Bhutan',
-      'BO': 'Bolivia',
-      'BA': 'Bosnia and Herzegovina',
-      'BW': 'Botswana',
-      'BV': 'Bouvet Island',
-      'BR': 'Brazil',
-      'VG': 'British Virgin Islands',
-      'IO': 'British Indian Ocean Territory',
-      'BN': 'Brunei Darussalam',
-      'BG': 'Bulgaria',
-      'BF': 'Burkina Faso',
-      'BI': 'Burundi',
-      'KH': 'Cambodia',
-      'CM': 'Cameroon',
-      'CA': 'Canada',
-      'CV': 'Cape Verde Islands',
-      'KY': 'Cayman Islands',
-      'CF': 'Central African Republic',
-      'TD': 'Chad Republic',
-      'CL': 'Chile',
-      'CN': 'China',
-      'HK': 'Hong Kong',
-      'MO': 'Macao',
-      'CO': 'Colombia',
-      'KM': 'Comoros',
-      'CG': 'Congo',
-      'CD': 'Dem. Republic of the Congo',
-      'CR': 'Costa Rica',
-      'HR': 'Croatia',
-      'CU': 'Cuba',
-      'CY': 'Cyprus',
-      'CZ': 'Czech Republic',
-      'DK': 'Denmark',
-      'DJ': 'Djibouti',
-      'DM': 'Dominica',
-      'DO': 'Dominican Republic',
-      'EC': 'Ecuador',
-      'EG': 'Egypt',
-      'SV': 'El Salvador',
-      'GQ': 'Equatorial Guinea',
-      'ER': 'Eritrea',
-      'EE': 'Estonia',
-      'ET': 'Ethiopia',
-      'FO': 'Faroe Islands',
-      'FJ': 'Fiji',
-      'FI': 'Finland',
-      'FR': 'France',
-      'GF': 'French Guiana',
-      'PF': 'French Polynesia',
-      'GA': 'Gabon Republic',
-      'GM': 'Gambia',
-      'GE': 'Georgia',
-      'DE': 'Germany',
-      'GH': 'Ghana',
-      'GR': 'Greece',
-      'GL': 'Greenland',
-      'GD': 'Grenada',
-      'GP': 'Guadeloupe',
-      'GU': 'Guam',
-      'GT': 'Guatemala',
-      'GG': 'Guernsey',
-      'GN': 'Guinea',
-      'GW': 'Guinea-Bissau',
-      'GY': 'Guyana',
-      'HT': 'Haiti',
-      'HN': 'Honduras',
-      'HU': 'Hungary',
-      'IS': 'Iceland',
-      'IN': 'India',
-      'ID': 'Indonesia',
-      'IR': 'Iran',
-      'IQ': 'Iraq',
-      'IE': 'Ireland',
-      'IM': 'Isle of Man',
-      'IL': 'Israel',
-      'IT': 'Italy',
-      'JM': 'Jamaica',
-      'JP': 'Japan',
-      'JE': 'Jersey',
-      'JO': 'Jordan',
-      'KZ': 'Kazakhstan',
-      'KE': 'Kenya',
-      'KI': 'Kiribati',
-      'KP': 'North Korea',
-      'KR': 'South Korea',
-      'KW': 'Kuwait',
-      'KG': 'Kyrgyzstan',
-      'LA': 'Lao PDR',
-      'LV': 'Latvia',
-      'LB': 'Lebanon',
-      'LS': 'Lesotho',
-      'LR': 'Liberia',
-      'LY': 'Libya',
-      'LI': 'Liechtenstein',
-      'LT': 'Lithuania',
-      'LU': 'Luxembourg',
-      'MK': 'Macedonia',
-      'MG': 'Madagascar',
-      'MW': 'Malawi',
-      'MY': 'Malaysia',
-      'MV': 'Maldives',
-      'ML': 'Mali',
-      'MT': 'Malta',
-      'MH': 'Marshall Islands',
-      'MQ': 'Martinique',
-      'MR': 'Mauritania',
-      'MU': 'Mauritius',
-      'YT': 'Mayotte',
-      'MX': 'Mexico',
-      'FM': 'Micronesia',
-      'MD': 'Moldova',
-      'MC': 'Monaco',
-      'MN': 'Mongolia',
-      'ME': 'Montenegro',
-      'MS': 'Montserrat',
-      'MA': 'Morocco',
-      'MZ': 'Mozambique',
-      'MM': 'Myanmar',
-      'NA': 'Namibia',
-      'NR': 'Nauru',
-      'NP': 'Nepal',
-      'NL': 'Netherlands',
-      'AN': 'Netherlands Antilles',
-      'NC': 'New Caledonia',
-      'NZ': 'New Zealand',
-      'NI': 'Nicaragua',
-      'NE': 'Niger',
-      'NG': 'Nigeria',
-      'NO': 'Norway',
-      'OM': 'Oman',
-      'PK': 'Pakistan',
-      'PW': 'Palau',
-      'PS': 'Palestinian Territory',
-      'PA': 'Panama',
-      'PG': 'Papua New Guinea',
-      'PY': 'Paraguay',
-      'PE': 'Peru',
-      'PH': 'Philippines',
-      'PN': 'Pitcairn',
-      'PL': 'Poland',
-      'PT': 'Portugal',
-      'PR': 'Puerto Rico',
-      'QA': 'Qatar',
-      'RE': 'Réunion',
-      'RO': 'Romania',
-      'RU': 'Russian Federation',
-      'RW': 'Rwanda',
-      'KN': 'Saint Kitts and Nevis',
-      'LC': 'Saint Lucia',
-      'VC': 'Saint Vincent and Grenadines',
-      'WS': 'Samoa',
-      'SM': 'San Marino',
-      'ST': 'Sao Tome and Principe',
-      'SA': 'Saudi Arabia',
-      'SN': 'Senegal',
-      'RS': 'Serbia',
-      'SC': 'Seychelles',
-      'SL': 'Sierra Leone',
-      'SG': 'Singapore',
-      'SK': 'Slovakia',
-      'SI': 'Slovenia',
-      'SB': 'Solomon Islands',
-      'SO': 'Somalia',
-      'SS': 'South Sudan',
-      'ES': 'Spain',
-      'LK': 'Sri Lanka',
-      'SD': 'Sudan',
-      'SR': 'Suriname',
-      'SZ': 'Swaziland',
-      'SE': 'Sweden',
-      'CH': 'Switzerland',
-      'SY': 'Syrian Arab Republic',
-      'TW': 'Taiwan',
-      'TJ': 'Tajikistan',
-      'TZ': 'Tanzania',
-      'TH': 'Thailand',
-      'TL': 'Timor-Leste',
-      'TG': 'Togo',
-      'TO': 'Tonga',
-      'TT': 'Trinidad and Tobago',
-      'TN': 'Tunisia',
-      'TR': 'Turkey',
-      'TM': 'Turkmenistan',
-      'TV': 'Tuvalu',
-      'UG': 'Uganda',
-      'UA': 'Ukraine',
-      'AE': 'United Arab Emirates',
-      'GB': 'United Kingdom',
-      'US': 'United States of America',
-      'UY': 'Uruguay',
-      'UZ': 'Uzbekistan',
-      'VU': 'Vanuatu',
-      'VE': 'Venezuela ',
-      'VN': 'Viet Nam',
-      'YE': 'Yemen',
-      'ZM': 'Zambia',
-      'ZW': 'Zimbabwe'
-    }
 
-    let toRet = []
-    for (var key in tmp) {
-      let el = {}
-      el.id = key
-      el.name = tmp[key]
-      toRet.push(el)
-    }
-    return toRet
-  }
 }
